+++ src/main/kotlin/dev/aaronhowser/mods/geneticsresequenced/event/ModBusEvents.kt
@@ -95,10 +95,6 @@ object ModBusEvents {
         if (!event.has(EntityType.PLAYER, ModAttributes.EFFICIENCY)) {
             event.add(EntityType.PLAYER, ModAttributes.EFFICIENCY)
         }
-
-        if (!event.has(EntityType.PLAYER, ModAttributes.WALL_CLIMBING)) {
-            event.add(EntityType.PLAYER, ModAttributes.WALL_CLIMBING)
-        }
     }
 
 }